player/requirements.txt
b2sdk>=1.23.0
python-dotenv>=1.0.0
Pillow>=10.0.0

